Removed fancy config writing and ruamel.yaml dependency
