+++ auth.yaml
@@ -4,12 +4,13 @@ version: 0
 faucet:
     prometheus_port: 9244
     ip: 127.0.0.1
+
 files:
     # the location files, pid should contain the process id (pid) of the main faucet process (ryu-manager)
     #  should be obtainable with "ps aux |grep faucet.faucet |head -n1 | awk '{print $2}' > /etc/ryu/contr_pid"
     controller_pid: /etc/ryu/contr_pid
     faucet_config: /etc/ryu/faucet/faucet.yaml
-    mac_learning: /faucet-src/faucet_mac_learning.txt
+    acl_config: /etc/ryu/faucet/acls.yaml
 
 urls:
     # HTTP endpoints for HTTPServer.py
