+++ faucet/config_parser.py
@@ -21,6 +21,8 @@ from vlan import VLAN
 from router import Router
 from watcher_conf import WatcherConf
 
+import yaml
+
 import config_parser_util
 
 
@@ -203,3 +205,106 @@ def _watcher_parser_v2(conf, logname):
             result.append(watcher)
 
     return result
+
+
+def write_acls_file(acls, filename):
+    """Overwrites the specified (in data) configurations.
+    :param top_level the name of the top level dict e.g. 'acls', 'dps', ...
+    :param data list of LocusCommentedMap object to overwrite part of the current config.
+        The top map object specifies the config_file.
+    """
+
+    # TODO this assumes that the top level structure wont be changed and only a sub structure.
+    # might want to add so all acls can be written, etc. and that the 2nd level structures do not
+    # share the same name with another 2nd level, even if different parent.
+
+
+    # 'i' will be a LocusCommentMap with the acl or dp. i.conf_file is the file for all its children. 
+    # for each second level structure
+
+        # dump yaml
+    yaml.dump(acls, open(filename, 'w'), default_flow_style=False) 
+
+
+def load_acl(config_path, switchname, switchport):
+    """Loads the acl as specified by the acl_in field on switchname switchport.
+    Both of switchname and switch must be specified.
+    :param config_path path to yaml configuration file to load.
+    :param switchname name of switch/datapath.
+    :param switchport the port of switchname.
+    :return tuple of dp name and dict-like config object.
+    """
+ 
+    top_conf = load_top_conf(config_path)
+    
+    acls = top_conf["acls"]
+    dps = top_conf["dps"]
+
+    acl_in = dps[switchname]["interfaces"][switchport]["acl_in"]
+
+    acl_map = acls[acl_in]
+    if acl_map is not None:
+        print("acl_in {}, acl_map {}".format(acl_in, acl_map))
+        return acl_in, acl_map
+
+    raise NotInYAMLError("Cannot find acl with dp named: {}, and port: {} in config file {}".format(switchname, switchport, config_path))
+
+
+def load_acls(config_path):
+    """Loads all acls across all files in config_path (include/include-optional)
+    :param config_path path to yaml configuration file to load.
+    :return dict of <name, LocusCommentedMap>
+    """
+    return yaml.load(open(config_path,'r'))
+#    return load_top_conf(config_path)["acls"]
+    
+
+def load_dp(config_path, switchname=None, dp_id=None):
+    """Loads a single datapath.
+    One of switchname or dp_id must be specified.
+    :param config_path path to yaml configuration file to load.
+    :param switchname name of switch/datapath to search for.
+    :param dp_id id of switch/datapath to search for.
+    :return tuple of dp name and dict-like config object.
+    """
+    dps = load_dps(config_path)
+    if switchname is not None:
+        return switchname, dps[switchname]
+
+    for name, com_map in dps.items():
+        if dp_id is not None and com_map["dp_id"] == dp_id:
+            return name, com_map
+
+    raise NotInYAMLError("Cannot find dp named: {}, or dp id: {} in config file {}".format(switchname, dp_id, config_path))
+
+
+def load_dps(config_path):
+    """Loads all datapaths across all files in config_path (include/include-optional)
+    :param config_path path to yaml configuration file to load.
+    :return dict of <name, LocusCommentedMap> e.g. s1 : LCM (dpid:10000, ...)
+    """
+    top = load_top_conf(config_path)
+    return top["dps"]
+
+
+def load_top_conf(config_path):
+    """Loads the top level configuraions.
+    :param config_path path to yaml configuration file.
+    :return dict of 4 main top level config LocusCommentedMap
+    """
+    config_hashes = {} 
+    top_confs = {
+            "acls": {},
+            "dps": {},
+            "routers": {},
+            "vlans": {}
+            }
+    config_parser_util.dp_include(config_hashes, config_path, "loadtop", top_confs)
+
+    return top_confs
+
+
+class NotInYAMLError(LookupError):
+    """Exception for if object cannot be found in the loaded yaml file object.
+    """
+    pass
